TSK-ID-13 Fix Navigation * fix CustomNavTypeSerializer for creating correct routes for the screens * fix navigation icon color on selected state
